move auto shuttle call files to appropriate place
