--- Content.Server/GameTicking/Rules/Components/EmergencyShuttleAutoCallRuleComponent.cs
@@ -1,8 +1,0 @@
-namespace Content.Server.GameTicking.Rules.Components
-{
-    [RegisterComponent, Access(typeof(EmergencyShuttleAutoCallRuleSystem))]
-    public sealed partial class EmergencyShuttleAutoCallRuleComponent : Component
-    {
-        //sad and empty... for now
-    }
-}
