+++ Content.Server/SS220/GameTicking/Rules/Components/EmergencyShuttleAutoCallRuleComponent.cs
@@ -1,0 +1,11 @@
+// © SS220, An EULA/CLA with a hosting restriction, full text: https://raw.githubusercontent.com/SerbiaStrong-220/space-station-14/master/CLA.txt
+
+using Content.Server.SS220.GameTicking.Rules;
+
+namespace Content.Server.GameTicking.Rules.Components;
+
+[RegisterComponent, Access(typeof(EmergencyShuttleAutoCallRuleSystem))]
+public sealed partial class EmergencyShuttleAutoCallRuleComponent : Component
+{
+    //sad and empty... for now
+}
